feat: Add @radix-ui/react-progress dependency and update page layouts
This commit adds the "@radix-ui/react-progress" dependency to the project by updating the "package.json" and "package-lock.json" files. It also updates the page layouts in the "page.tsx" and "layout.tsx" files to include the new progress component. These changes enhance the user experience by providing a visual progress indicator on the page.
Along with these changes, the navbar for the homepage was created.

Refs: #123