refactor: Update Comboio component to use conditional rendering for image link
This commit refactors the Comboio component in the homepage to use conditional rendering for the image link. Instead of always rendering a link to `/chapter/${carruagem.id}/${carruagem.id}`, the link is now conditionally rendered based on the value of `carruagem.id`. If `carruagem.id` is 0, the image is displayed without a link. Otherwise, the link is rendered as `/chapters/chapter${carruagem.id}/1`. This change improves the navigation and user experience of the homepage.

Refs: #123