refactor: Update background images in Chapter3Page1 and Chapter4Page1 components
This commit updates the background images in the Chapter3Page1 and Chapter4Page1 components. The background image in Chapter3Page1 is changed to '/img/fundo_comunicacao.svg', while the background image in Chapter4Page1 is changed to '/img/fundo_tecnologia.svg'. These changes improve the visual appeal and consistency of the application.

Refs: #123