refactor: Update Chapter4Page2 and Chapter4Page5 components
The Chapter4Page2 component has been updated to fix a bug where the slider value was not being properly handled. The handleSlideChange function has been modified to correctly update the slider value.

The Chapter4Page5 component has been updated to check if the currentUser object is not null before calling the goGetUsername function. This prevents an error from occurring when the currentUser is null.

Refs: #123